cloud_backend/api/app_script/requirements.txt
paho-mqtt>=1.6.1
requests>=2.28.1
ultralytics>=8.0.0
opencv-python>=4.6.0
numpy>=1.23.5
Pillow>=9.3.0
